subida ok

+++ spring-boot/src/main/java/com/lamarrulla/cargaCSV/springboot/dto/FooterDTO.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 public class FooterDTO {
     private static final Long serialVersionUID = 1L;
     private Long idFooter;
-    private Long luno;
-    private Long ldos;
-    private Long total;
+    private Double luno;
+    private Double ldos;
+    private Double total;
 }
